requirements.txt
neo4j==5.15.0
typing_extensions==4.9.0
pyyaml==6.0.1
sanic==23.6.0
